remove MainWindow and Parameters widgets, and split remaining code in aperoll.widgets.parameters into a new module aperoll.widgets.line_edit and aperoll.utils
